+++ BoomerBorrow/app.json
@@ -38,7 +38,8 @@
         {
           "RNMapboxMapsDownloadToken": "sk.eyJ1Ijoicm9zbzQ3ODUiLCJhIjoiY205aGh6bTJ3MDRmZjJrcjE2eGVraGR2dyJ9.6AuHtcQWsGIcNqquPh9lkg"
         }
-      ]
+      ],
+      "expo-secure-store"
     ],
     "experiments": {
       "typedRoutes": true
