+++ BoomerBorrow/app/(tabs)/(login)/index.tsx
@@ -4,7 +4,7 @@ import { Text, TextInput, Button, StyleSheet } from "react-native";
 import { router } from "expo-router";
 import axios from "axios";
 import { User } from "./../../../classes_tmp";
-import token_storage from "@/token_storage";
+import { save_user_id, get_user_id } from "@/auth_token";
 
 export default function LoginScreen() {
 	const SERVER_URL = "http://localhost:3000";
@@ -36,8 +36,9 @@ export default function LoginScreen() {
 				password: password
 			}
 			const response = await axios.post(`${SERVER_URL}/login`, user);
-			token_storage.store_token(response.data);
-			console.log(`Stored auth_token: ${response.data}`);
+			await save_user_id(response.data);
+			
+			console.log(`Retrieved user id: "${response.data}"`);
 
 			// Sends the client to the marketplace page
 			router.push("/(tabs)/(marketplace)");
