+++ BoomerBorrow/app/(tabs)/(marketplace)/index.tsx
@@ -8,7 +8,8 @@ import Mapbox, { MapView, Camera } from "@rnmapbox/maps";
 // Set the Mapbox access token
 Mapbox.setAccessToken("pk.eyJ1Ijoicm9zbzQ3ODUiLCJhIjoiY205Z3Q4azlpMXN6cTJrcXc3anNhN2d2eCJ9.gYQgEn_h2O1CGIxWkEpcdA");
 import { Supply_post } from "./../../../classes_tmp";
-import token_storage from "@/token_storage";
+import { get_user_id } from "@/auth_token";
+//import token_storage from "@/token_storage";
 
 export default function MarketplaceScreen() {
   const SERVER_URL = "http://localhost:3000";
@@ -44,23 +45,11 @@ export default function MarketplaceScreen() {
       const created_at = new Date().toISOString();
 	
 	  // TODO: need unique ID:s for every post
-      supply_post = new Supply_post(
-        "20",    // ID
-        "10",    // Owner ID
-        title,   // title
-        description,    // description
-        price,    // price
-        category,   // category
-        location,   // location
-        post_picture_url,   // post_picture_url
-        created_at    // created at
-      );
-
-      const headers = {
-        "Authorization": `${token_storage.get_token}` // Attach token here
-      };
-
-      await axios.post(`${SERVER_URL}/new_supply_post`, supply_post, { headers });
+      const owner_id = await get_user_id();
+      if (owner_id === null ) { throw new Error("Owner id cannot be null upon post creation"); }
+      supply_post = new Supply_post(owner_id, title, description, price, category, location, post_picture_url, created_at);
+      
+      await axios.post(`${SERVER_URL}/new_supply_post`, supply_post, { headers: { auth: `${await get_user_id()}` } });
       console.log("post_data (the new post) sent to the server");
 
       const response = await axios.get(`${SERVER_URL}/fetch_all_supply_posts`);
@@ -90,7 +79,7 @@ export default function MarketplaceScreen() {
 
   async function delete_supply_post(post_id: string) {
     try {
-      const response = await axios.post(`${SERVER_URL}/delete_supply_post`, { title: post_id });
+      const response = await axios.post(`${SERVER_URL}/delete_supply_post`, { id: post_id });
       
       if (response.status === 200) {
         await fetch_active_supply_posts();
@@ -150,8 +139,8 @@ export default function MarketplaceScreen() {
               <Text style={{ fontWeight: "bold" }}>{supply_post.title}</Text>
               <Text>{supply_post.description}</Text>
 			        <Text>{supply_post.price}</Text>
-			        <Text>{supply_post.category}</Text>                   {/*   // TODO: post_title should be post_id */}
-              <TouchableOpacity style={styles.minibutton} onPress={() => handle_supply_post_deletion(supply_post.title)}>
+			        <Text>{supply_post.category}</Text>
+              <TouchableOpacity style={styles.minibutton} onPress={() => handle_supply_post_deletion(supply_post.id)}>
               <Text>Radera</Text>
               </TouchableOpacity>
             </View>
