+++ BoomerBorrow/package.json
@@ -31,6 +31,7 @@
     "expo-haptics": "~14.0.1",
     "expo-linking": "~7.0.5",
     "expo-router": "^4.0.20",
+    "expo-secure-store": "~14.0.1",
     "expo-splash-screen": "~0.29.22",
     "expo-status-bar": "~2.0.1",
     "expo-symbols": "~0.2.2",
@@ -53,6 +54,7 @@
     "@types/jest": "^29.5.12",
     "@types/react": "~18.3.12",
     "@types/react-test-renderer": "^18.3.0",
+    "@types/uuid": "^10.0.0",
     "@types/ws": "^8.18.1",
     "babel-plugin-module-resolver": "^5.0.2",
     "jest": "^29.2.1",
